Add plugin and process for putting content into search
The jekyll_rummageable plugin will parse the HTML generated during the
compilation process and create a JSON version of it suitable for pushing
into rummager.

Changes to compile.sh copy that JSON into designprinciples ready to be
part of its deploy process.

This has been tested end-to-end in a dev VM but should be reviewed on
preview. It may well be that we also want to amend the plugin to use the
raw markdown documents for generating the JSON rather than HTML, but this
approach was the simplest to implement for now.