Strengthen guidance on accessibility and content
A number of wording tweaks. In general I've made the guidance on accessible
formats and open standards stronger, and made clear that Léonie's guidance on
making PDFs accessible is help for a "last resort" situation, and not our
desired publication method.

In addition, I've added the cross-referencing chatted about with @psd.