This sentence is a bit unwieldy
Not sure whether the suggestion I've made is the right one, but something could be done with commas or hyphens to make it a bit more readable? I know it's a quote from the report, but you've changed the & to 'and' so a couple more small changes might not go awry?