Fix reference error that was preventing generation of tables of contents
Because we call render on page objects we were creating a copy of our list
of page objects before rendering them to extract content for search. In
practice this wasn't working because we were copying the list but it still
gave us a list of references to the same page objects. Instead we need to
copy each page object individually and render the copy.