Use the first paragraph as search summary
Previously we'd take all content and extract the first text item from
it, which was almost always the table of contents. This results in
summaries that match the H2s of the document and was generally useless.

This instead takes the first paragraph as the search summary.

Worked on this with @alextea.