Tried to make it a bit easier to understand and some minor corrections and additions
The first sentence in paragraph 3 of the Objectives section was a bit hard to read and understand, because of the "user need analysis needs" phrase. I tried to simplify it. Also added a missing "." to the "i.e." and a link to the "Assisted Digital" since I didn't understand what "potential assisted digital needs" really meant before I had read that page. This manual really is great though. I hope these suggestions are helpful.