Use header-footer only CSS and add relevant styles
We were relying on inheriting styling from static, which is dangerous as it could change without notice and break the styling of the manual (as seen with search alphagov/design-principles#38).

This means we are now only including the CSS for the header and footer and defining our own styles based on static.