First pass on new domain name/service.gov.uk guidance
We're getting lots of enquiries about the details of getting and running a
service.gov.uk domain. This is a first pass at restructuring our guidance
about those domains to make some details clearer, flesh out some points and
to begin to make it clear that the approvals are part of a wider process.

So far this is very sketchy and intended primarily to move an existing
conversation out of emails/google docs and into 'code'