Update Git submodules before compiling
The GOV.UK frontend toolkit is included as a submodule, which isn't
documented. Without this Sass compilation fails, which causes all the
assets to be destroyed. Unfortunately the asset compilation exits with 0
so there's no easy way to tell it failed.

This forces a submodule update before trying to run Jekyll so that we
always have the desired version of external things.