Put in GDS style, removed marmite metaphor.
Define 'acceptance criteria'. This is the first (of many) mentions of acceptance criteria in the agile docs. An explanation here is essential.

"user story card" needs explaining - link it to the 'writing user stories' document

Can we delete this para? "To stop your meetings from becoming repetitive, have an object that you (gently) throw to someone to signify they should speak next. Pick people at random - it keeps people on their toes and lets the person speaking to choose the person they wish to hear from next. At GDS we use cuddly toys or a piece of fruit. It’s a bit of fun. You don’t have to this - it’s just something to experiment with." If it's something a few GDS teams do rather than agile methodology it's a bit cutesy for this document.

The marmite metaphor doesn't work - although the properties of Marmite might be good for you, they aren't exclusive to Marmite and can be provided by other products. Using an advertising slogan, now so old it's passed into cliche, is a big change in tone - from the authoritative to the avuncular - for this document.