Removed para about using cuddly toys and fruit as props in meetings.
Removed this para as it deviates from the authoritative tone. Some users may not find this puts GDS in a suitable light: 

To stop your meetings from becoming repetitive, have an object that you (gently) throw to someone to signify they should speak next. Pick people at random - it keeps people on their toes and lets the person speaking to choose the person they wish to hear from next. At GDS we use cuddly toys or a piece of fruit. It’s a bit of fun. You don’t have to this - it’s just something to experiment with.