Changed FAQ H3s, clarified points; more clarification required.
Clarified first sentence. 

"A transaction completed only partially online means the user may complete a single process within the overall transaction, like booking an appointment or completing a part of an application." What does this mean in terms of measurement? Will it be measured, or does it fall outside of the end to end process? Please clarify.