style guide compatibility
high-minded can mean arrogant; it usually means principled, so 'high-minded principles' will be tautological to most people and off-putting to others.
Bulleted lists are for points of one line/sentence maximum, so changed the Good Habits to a numbered list. Also modified its lead-in so it made sense that it was about the points below, not the principles above.