Switch to using jekyll-assets to include assets
This commit uses the jekyll-assets gem to generate nice asset paths
that are cachebusted properly.

Rename all the JavaScript and Stylesheets to put them in the _assets
directory. Put the toolkit submodule there too.

Remove the CSS check from the compile script. Compilation definitely
fails now if the Sass doesn't compile, as I discovered when it failed
because we have a stray image-url() in _typography.scss (removed).

Use the provided liquid tags to include the CSS and JS in the HTML
document.