Fix page.title when using govuk_template
govuk_template only allows us to include page.title in the title tag,
where we also want a suffix to indicate that this is the Service Manual.

However, we also make use of page.title inside h1 tags.

This commit adds a plugin that sets the title to include the pretty suffix,
sets an attribute on the data that doesn't include the suffix and then converts
all of the Service Manual specific layouts to use the new un-suffixed title.