Update privacy-note-template-for-services
I've spoken at length to the content team: "it's been quite a struggle to get it to this stage. It's mainly for lawyers and we've checked it with quite a few of them already. None of them had any problems with how it would be used."

I've agreed with them that we should add an introductory explanation above the subhead.