Caveat about using the same DNS provider as GDS
I wasn't sure whether this is appropriate for the service manual, as it
references a particular provider (I have avoided hyperlinking to avoid
the idea that we are promoting them over any other).

We have an issue when Transactional Services want to use the same DNS
provider as GDS uses to manage the service.gov.uk domain. In order for
the backend systems to know which customer account is authorative for
that particular record, the provider needs to make some alterations and
requires a signed letter of authorisation from both parties.

As this is outside of our control and may affect timescales, I thought
it worth documenting here.