Quick prototype of creating a book format
We have been asked for a pdf version of the service manual in the past.
I’m not completely sure what the user need is for this, but here is a
very basic sketch around that.

To try it out, you’ll need to install a few things:

* `sudo aptitude install node`
* `npm install`

If that works, you can then create different versions of the content:

```
$ gitbook pdf ./
```