Cleaning up the Resources for developers section
Different people use the tags differently, with some pages having
_everyone_ as the primary audience and others being more selective.
This to me has led to a resources page that doesn't really work for
anyone.

I've tried to take a view on what a developer might be
interested in on the main resources page, moving more specialist content
to the secondary list.

This is probably quite opinionated so suggestions welcome. I'm not
saying something isn't important, I'm saying some things are relevant to
everyone (version control, progressive enhancement, testing) and some
things (CAPCHAs, load testing, Sass) are relevant either to specialists
or occasionally or both.