Make elements page show up in search
Only exclude index.md files that have a role-index layout. That layout
means that the page is mainly intended to be a catalogue of links rather
than actual content. A page of links should not show up in our search
results, since that would introduce an extra click to the user journey.

After this change:

* https://www.gov.uk/service-manual/search?q=captcha
  should only have a single result.
  https://www.gov.uk/service-manual/developers should not be in the
  results for that search. This is unchanged behaviour.
* https://www.gov.uk/service-manual/search?q=colour should have
  https://www.gov.uk/service-manual/user-centred-design/resources/elements
  in the results for that search. This is the new, desired behaviour.