Link for published as OSS is incorrect
The link here for 'published as open source software' is to the alphagov organisation on GitHub, but this [explicitly isn't OSS](https://gds.blog.gov.uk/2012/10/12/coding-in-the-open/). I have replaced this with a link to the gds-operations organisation which [is OSS](https://github.com/gds-operations/open-source-guidelines), though there might be a better link, such as [Gareth's PR on OSS on the service manual](https://github.com/alphagov/government-service-design-manual/pull/360), when published.