Add spellchecking tool
This is similar to what's currently in the Makefile, but it does a little
bit more.

It's quite flaky, like everything I put in the tools directory. But
that's ok because it doesn't run anywhere important and it finds
typos, which is better than not finding typos.

The number on the 1st line of the dictionary file is the number of
words in the dictionary. I think it's approximate so that aspell
knows what kind of size of memory it's going to be dealing with.