A guide to help understand the scope of web operations work
This is based on a Google document I wrote about a year ago that has
seen extensive use on a number of exemplar projects and that is being
shared a little too widely.

I'm stil skeptical that anything that appears to be this prescriptive is
a good idea, but in practice this has worked very well and had the
desired effect, namely to make people much more aware of the amount of
work required. So it's well tested, if on the face of it a bit crazy.

For work like this that is broadly similar for many services starting
with many things and then adding a few and removing a few appears better
than starting from scratch everywhere.