Exclude vendor directory from Jekyll
When using this app with Jekyll 1.2.1, `jekyll serve --watch` complains that
the site could not be built because
`0000-00-00-welcome-to-jekyll.markdown.erb` does not have a valid date, and
therefore is not syntactically valid.

This fix adds the `vendor` directory to Jekyll's exclude list in
`_config.yml`. Jekyll serving then works fine.