Upgrade dependencies
We’d like to be able to run this on heroku, so that the team
responsible for the governance project can publish their changes and do
user research before putting their content live.

There are various issues with our current dependency versions which are
preventing that.

Upgrading to latest versions of things and then we’ll make pull requests
if bugs / features are still an issue.