Correct DbD wording
The content that was previously linked to was removed in 0178f4ea0d35a6489586a16944e548bbaa7d15a4. This change also amended the
service standard wording as a result of removing the linked content.

We are not supposed to change wording of the service standard. Any
changes need to be cleared with the service standard team.

Added the words back, and now linking to the content style guide for
GOV.UK.