Add search to the service manual
This includes a bit of duplication from the regular Jekyll service
manual, but it's all contained in the _search directory. As explained
in that README, I believe this is due to the way Jekyll implements
Liquid.

Bump the version of GDS API adapters because the version I included
from the design-principles app is out of date.