Add link checking for internal product links
I would like to ensure that what we publish doesn’t contain broken
links. This is basic quality check bashed together on a couple of
commutes and an evening.

It found a problem, so I’m happy.

It is slightly hardcoded to the service-manual product. That is
intentional – this is a short-term hacky MVP to try to raise quality.