Added a link to our free eBook with Retrospective Exercises
I think our book can be a valid resource for people working in government that want to agile retrospectives. Therefor I added a link to this page. Let me know if that is ok or not ok?

We have decided to make our book available through Leanpub for a minimum price of $0.00 (free) to help teams all around the world to get more value from doing agile retrospectives.