bullets-in-assisted-digital-main-page
Need to fix the formatting of lines 76-78 and 95-97 - should be
bulleted, but are appearing as asterisks. I’ve copied it out, pasted
back in a cleaned version, but Github’s not picking these up in the
commit as they’re no different to before. Please have a look and see
what’s missing to make them bullets?
thank you
Ben