Index documents using the "manual" formats
This depends on https://github.com/alphagov/rummager/pull/399; I'll
deploy both that and this change together.

So that we can move to using the `unified_search` endpoint, we want to
index documents from the service manual the same as if they were from
one of the standard manuals on GOV.UK (ie, using the recently introduced
`manual` and `manual-section` formats).

This requires several changes:

 - the documents from the manual need to be indexed with their type set
   to `manual_section`.

 - manuals are generally associated with an organisation, so associate
   this manual with the GDS organisation.

 - the titles in search of all existing `manual_section` documents are
   prefixed with the name of the manual.  This makes them make more
   sense when they come up in general site search, but is undesirable
   when the search is restricted to a single manual.  Therefore, we also
   strip this prefix off again when presenting the results in this app.

 - for search-within-a-manual to work fully, we need a `manual` document
   to describe the manual.  This is a good idea anyway, because
   we're currently using the "recommended-links" app to add a link to
   search, so that a search for "service-manual" on GOV.UK finds it.
   This document requires a "description" field, for which I've copied
   the existing description used for the link in the "recommended-links"
   app.

This commit shouldn't change the results shown to users at all.

When deploying this commit, the old service-manual search index should
first be manually wiped. This is because the types of the documents have
changed, and therefore the old documents won't be removed or replaced in
the index.