Update the version of Rummageable in use
This is a major upgrade; the API of rummageable was completely rewritten
(nearly 2 years ago now!).  Fortunately, it's very easy to update the
code to use this, since it's only a few lines in the Rakefile.

This upgrade needs to be done now because the old version of rummageable
doesn't permit sending most fields, and I want to start sending
appropriate fields to make the documents indexed here be the same as
documents from other manuals.