Use the unified search endpoint
Instead of using the old /search endpoint on rummager, use the
/unified_search endpoint, applying a filter on the manual field to
return only service-manual documents.

In my testing, this returns the same results as using the /search
endpoint, and they look the same.  The ordering may vary slightly,
because the unified_search endpoint applies some popularity adjustments
to its weighting, but I expect these to be an improvement, if anything.

Also, use Plek.find("rummager") instead of Plek.find("search") now that
we've defined a rummager name; we should start moving to that, instead
of search when we're making changes in that area.