Change wording about DKIM/SPF
SPF provides a public list of servers that you expect to be sending mail for
your domain. The recipient can check this list to determine whether the
server from which it has recieved mail is on the list and hence whether the
message is likely to be authentic. This method can be spoofed by careful editing
of the email Received-From header, so it is an indicator of authenticity, not
a guarantee.

DKIM (DomainKey Identified Mail) allows you to publish a public key in DNS and
to sign all your mail with your private key. The recipient can check that any
mail it receives are signed with that public key and hence can know that the
email it has received has been sent by the organisation in control of that
domain name. It is hard to forge unless you are in control of both the sending
server (and private key) and the domain name, but can be spoofed.

Neither of these two methods provide guarantees that a message will be delivered
successfully to the recipients mailbox. Some mail systems will tend to assign
higher degrees of confidence that messages are not spam to emails signed with
DKIM and from mailservers published in an SPF record, but it's still possible
for these messages to be flagged as spam due to the content or other factors.

This changes a single word - delivery - to authenticity. By providing more
information about the authenticity of our messages, we allow the destination to
make better choices about how to deliver flag mail, but we still do not get any
guarantees about delivery.