Properly capitalise DevOps
DevOps is a portmanteau of 'development' and 'operations', and in some ways
is a good reflection of one of its core tenets. Since each word can be used
individually, we should style this correctly.

Colloboration amongs operations and development teams is central to DevOps,
so since neither party rules the roost all the time, we should capitalise
the start of each part of the portmanteau.