Update mobile Safari entries in verified browsers
Since the last update, Safari 9 has appeared and
is the most-used version while Safari 8.x is
decreasing but is still being used by over 6% of
our users.

Safari 7.0 is on the edge of the 95%, currently
at 0.62%). Safari 7.1 is a lot higher at 1.66%
but this has steadily decreasing all year having
started at 8.37%. This should be revisited in the
next few months.

Safari 5 and 6 are both now around 0.019% and
steadily decreasing.