Document the exclude_from_search metadata
As we migrate sections away from the old service manual we need to stop indexing the old service manual in search. The functionality already exists.

It seems it's possible to add an array of patterns to exclude from search in config.yml too. This might be useful too but we think for the moment that it's safer and easier to manage if we handle the exclusion in the metadata.